+++ resources/views/markdown/changelog.md
@@ -3,6 +3,12 @@
 Last Update: June 4, 2024\
 Travis Michael Cross <a href="mailto:travis@not.tv">travis@not.tv</a>
 
+## v0.9.3.16
+June 5, 2024
+
+* Refactor how recordings are saved to generate download and share urls.
+* Created a UpdateRecordingUrls command to update old recordings.
+
 ## v0.9.3.15
 June 5, 2024